requirements.txt
beautifulsoup4
requests
python-nmap
python-docx
pdfkit
argparse
urllib3
colorama
playwright
sublist3r
